split up cli, client and orm into smaller modules
