+++ src/maestral/cli/__init__.py
@@ -1,0 +1,3 @@
+from .cli_main import main
+
+__all__ = ["main"]
