+++ src/maestral/daemon.py
@@ -28,7 +28,7 @@ import sdnotify
 from fasteners import InterProcessLock
 
 # local imports
-from .errors import SYNC_ERRORS, GENERAL_ERRORS, MaestralApiError
+from .exceptions import SYNC_ERRORS, GENERAL_ERRORS, MaestralApiError
 from .utils import exc_info_tuple
 from .utils.appdirs import get_runtime_path
 from .constants import IS_MACOS, ENV
@@ -120,12 +120,12 @@ def serpent_deserialize_api_error(class_name: str, d: dict) -> MaestralApiError:
     :returns: Class instance.
     """
 
-    # Import maestral.errors for evaluation.
+    # Import maestral.exceptions for evaluation.
     # This import needs to be absolute to reconstruct the Exception class. Note that the
     # eval is safe here because ``serpent_deserialize_api_error`` is only registered for
     # strings that match an error class name.
 
-    import maestral.errors  # noqa: F401
+    import maestral.exceptions  # noqa: F401
 
     cls = eval(class_name)
     err = cls(*d["args"])
@@ -595,31 +595,31 @@ def stop_maestral_daemon_process(
 class MaestralProxy:
     """A Proxy to the Maestral daemon
 
-    All methods and properties of Maestral's public API are accessible and calls /
-    access will be forwarded to the corresponding Maestral instance. This class can be
-    used as a context manager to close the connection to the daemon on exit.
+        All methods and properties of Maestral's public API are accessible and calls /
+        access will be forwarded to the corresponding Maestral instance. This class can be
+        used as a context manager to close the connection to the daemon on exit.
 
-    :Example:
+        :Example:
 
-        Use MaestralProxy as a context manager:
+            Use MaestralProxy as a context manager:
 
-        >>> with MaestralProxy() as m:
-        ...     print(m.status)
+    import src.maestral.cli.cli_info        >>> with MaestralProxy() as m:
+            ...     print(src.maestral.cli.cli_info.status)
 
-        Use MaestralProxy directly:
+            Use MaestralProxy directly:
 
-        >>> m = MaestralProxy()
-        >>> print(m.status)
-        >>> m._disconnect()
+    import src.maestral.cli.cli_info        >>> m = MaestralProxy()
+            >>> print(src.maestral.cli.cli_info.status)
+            >>> m._disconnect()
 
-    :ivar _is_fallback: Whether we are using an actual Maestral instance as fallback
-        instead of a Proxy.
+        :ivar _is_fallback: Whether we are using an actual Maestral instance as fallback
+            instead of a Proxy.
 
-    :param config_name: The name of the Maestral configuration to use.
-    :param fallback: If ``True``, a new instance of Maestral will be created in the
-        current process when the daemon is not running.
-    :raises CommunicationError: if the daemon is running but cannot be reached or if the
-        daemon is not running and ``fallback`` is ``False``.
+        :param config_name: The name of the Maestral configuration to use.
+        :param fallback: If ``True``, a new instance of Maestral will be created in the
+            current process when the daemon is not running.
+        :raises CommunicationError: if the daemon is running but cannot be reached or if the
+            daemon is not running and ``fallback`` is ``False``.
     """
 
     _m: Union["Maestral", Proxy]
