+++ src/maestral/database/orm.py
@@ -6,52 +6,34 @@ Maestral. Many operations will still require explicit SQL statements. This modul
 alternative to fully featured ORMs such as sqlalchemy but may be useful when system
 memory is constrained.
 """
-import os
+
 import sqlite3
-from enum import Enum
 from weakref import WeakValueDictionary
 from typing import (
     Union,
     Type,
-    Tuple,
-    Sequence,
     Any,
     Generator,
     List,
     Optional,
     TypeVar,
-    Iterable,
-    Iterator,
     FrozenSet,
 )
 
+from .query import Query
+from .types import SqlType, SqlEnum, ColumnValueType
 
-ColumnValueType = Union[str, int, float, Enum, None]
 DefaultColumnValueType = Union[ColumnValueType, Type["NoDefault"]]
 SQLSafeType = Union[str, int, float, None]
 T = TypeVar("T")
 
 
 __all__ = [
-    "SqlType",
-    "SqlString",
-    "SqlInt",
-    "SqlFloat",
-    "SqlPath",
-    "SqlEnum",
-    "Query",
-    "AllQuery",
-    "MatchQuery",
-    "PathTreeQuery",
-    "AndQuery",
-    "OrQuery",
-    "NotQuery",
     "Column",
     "NoDefault",
     "Database",
     "Manager",
     "Model",
-    "ColumnValueType",
 ]
 
 
@@ -61,198 +43,6 @@ class NoDefault:
 
     This is distinct from ``None`` which may be a valid default.
     """
-
-
-class SqlType:
-    """Base class to represent Python types in SQLite table"""
-
-    sql_type = "TEXT"
-    py_type: Type[ColumnValueType] = str
-
-    def sql_to_py(self, value):
-        """Converts the return value from sqlite3 to the target Python type."""
-        return value
-
-    def py_to_sql(self, value):
-        """Converts a Python value to a type accepted by sqlite3."""
-        return value
-
-
-class SqlString(SqlType):
-    """Class to represent Python strings in SQLite table"""
-
-    sql_type = "TEXT"
-    py_type = str
-
-
-class SqlInt(SqlType):
-    """Class to represent Python integers in SQLite table"""
-
-    sql_type = "INTEGER"
-    py_type = int
-
-
-class SqlFloat(SqlType):
-    """Class to represent Python floats in SQLite table"""
-
-    sql_type = "REAL"
-    py_type = float
-
-
-class SqlPath(SqlType):
-    """
-    Class to represent Python paths in SQLite table
-
-    This class contains special handling for strings with surrogate escape characters
-    which can appear in badly encoded file names.
-    """
-
-    sql_type = "BLOB"
-    py_type = str
-
-    def sql_to_py(self, value: Optional[bytes]) -> Optional[str]:
-        if value is None:
-            return value
-
-        return os.fsdecode(value)
-
-    def py_to_sql(self, value: Optional[str]) -> Optional[bytes]:
-        if value is None:
-            return value
-
-        return os.fsencode(value)
-
-
-class SqlEnum(SqlType):
-    """Class to represent Python enums in SQLite table"""
-
-    sql_type = "TEXT"
-    py_type = Enum
-
-    def __init__(self, enum: Iterable[Enum]) -> None:
-        self.enum_type = enum
-
-    def sql_to_py(self, value: Optional[str]) -> Optional[Enum]:  # type: ignore
-        if value is None:
-            return None
-
-        return getattr(self.enum_type, value)
-
-    def py_to_sql(self, value: Optional[Enum]) -> Optional[str]:
-        if value is None:
-            return None
-
-        return value.name
-
-
-class Query:
-    def clause(self) -> Tuple[str, Sequence[ColumnValueType]]:
-        raise NotImplementedError()
-
-
-class PathTreeQuery(Query):
-    def __init__(self, column: "Column", path: str):
-
-        if not isinstance(column.type, SqlPath):
-            raise ValueError("Only accepts columns with type SqlPath")
-
-        self.column = column
-        self.file_blob = os.fsencode(path)
-        self.dir_blob = os.path.join(self.file_blob, b"")
-
-    def clause(self):
-        query_part = f"({self.column.name} = ? OR substr({self.column.name}, 1, ?) = ?)"
-        args = (self.file_blob, len(self.dir_blob), self.dir_blob)
-
-        return query_part, args
-
-
-class MatchQuery(Query):
-    def __init__(self, column: "Column", value: ColumnValueType):
-        self.column = column
-        self.value = value
-
-    def clause(self):
-        args = (self.column.py_to_sql(self.value),)
-        return f"{self.column.name} = ?", args
-
-
-class AllQuery(Query):
-    def clause(self):
-        return "TRUE", ()
-
-
-class CollectionQuery(Query):
-    """An abstract query class that aggregates other queries. Can be
-    indexed like a list to access the sub-queries.
-    """
-
-    def __init__(self, *subqueries: Query):
-        self.subqueries = subqueries
-
-    # Act like a sequence.
-
-    def __len__(self) -> int:
-        return len(self.subqueries)
-
-    def __getitem__(self, key: int) -> Query:
-        return self.subqueries[key]
-
-    def __iter__(self) -> Iterator[Query]:
-        return iter(self.subqueries)
-
-    def __contains__(self, item: Query) -> bool:
-        return item in self.subqueries
-
-    def clause_with_joiner(self, joiner: str) -> Tuple[str, Sequence[ColumnValueType]]:
-        """Return a clause created by joining together the clauses of
-        all subqueries with the string joiner (padded by spaces).
-        """
-        clause_parts = []
-        subvals: List[ColumnValueType] = []
-
-        for subq in self.subqueries:
-            subq_clause, subq_subvals = subq.clause()
-            clause_parts.append("(" + subq_clause + ")")
-            subvals += subq_subvals
-
-        clause = (" " + joiner + " ").join(clause_parts)
-        return clause, subvals
-
-    def clause(self):
-        raise NotImplementedError()
-
-
-class AndQuery(CollectionQuery):
-    """A conjunction of a list of other queries."""
-
-    def clause(self):
-        return self.clause_with_joiner("AND")
-
-
-class OrQuery(CollectionQuery):
-    """A conjunction of a list of other queries."""
-
-    def clause(self):
-        return self.clause_with_joiner("OR")
-
-
-class NotQuery(Query):
-    """A query that matches the negation of its `subquery`, as a shorcut for
-    performing `not(subquery)` without using regular expressions.
-    """
-
-    def __init__(self, subquery: Query):
-        self.subquery = subquery
-
-    def clause(self):
-        clause, subvals = self.subquery.clause()
-        if clause:
-            return f"not ({clause})", subvals
-        else:
-            # If there is no clause, there is nothing to negate. All the logic
-            # is handled by match() for slow queries.
-            return clause, subvals
 
 
 class Column(property):
